Make event creation/editing possible through shared component
CreateEvent used to house the form to create events. When needing to
update forms I realized that we need the exact same form. By splitting
the form out to a separate component and adding an event handler to
specify what to do on submission now creating and editing can reuse this
component.