Switch to routify for frontend routing
This caused a lot of files to move since this router uses the filesystem
to build routes. This is an overall positive as the project was really
needing some structure. Also this router has a lot of helpers that is
maiking routing and protecting routes very nice. Future commits will
kill off the old `routes` folder once we can fully move over