Improve checking for admin users
There was a lot of fiddly code around checking if the user was an admin.
This new change makes the admin store derived off of the token store.
Anytime a new token is received the admin store can decode it and decide
if the user is an admin. This has the advantage of making the token the
  source of truth and not having to store admin information in
  localStorage

* Moved token store to be an observable store
* Make Admin store derived off of token store
* Update admin layout to use new admin store contract